cli/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "zen-ai-cli"
version = "1.0.2"
description = "A beautiful terminal interface for Zen AI"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    {name = "Zen AI Team", email = "your-email@example.com"}
]
keywords = ["cli", "ai", "terminal", "chat", "notes", "tui"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: End Users/Desktop",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Communications :: Chat",
    "Topic :: Utilities",
]
dependencies = [
    "httpx>=0.25.0",
    "rich>=13.0.0",
    "prompt_toolkit>=3.0.0",
    "python-dotenv>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/joan-code6/zen_ai"
Repository = "https://github.com/joan-code6/zen_ai"
Issues = "https://github.com/joan-code6/zen_ai/issues"

[project.scripts]
zen = "main:run"

[tool.setuptools]
py-modules = ["main", "api_client", "config", "ui", "selector"]

[tool.setuptools.packages.find]
where = ["."]
include = ["commands*"]
